Add autossh wrapper that automatically connects to same tmux
Generate a uuid on connect, if autossh reconnects before then, it will
use the same uuid and that is detected by the tmux chooser script.

That script will then iterate through all the sessions and connect to
the one that has the same uuid in the tmux session.

Note that this requires the LC_TMUX env variable to be sent via ssh,
which is default with debian based systems